kmeans/requirements.txt
pyspark==3.4.0
matplotlib==3.6.2
pandas==1.5.2
mqtt-client==1.6.0
psycopg2-binary==2.9.6